+++ rtloader/include/rtloader_types.h
@@ -144,7 +144,7 @@ typedef double (*cb_get_process_start_time_t)(void);
 // (cmd, error_message)
 typedef char *(*cb_obfuscate_mongodb_string_t)(char *, char **);
 // (check_name, metric_name, metric_value, metric_type)
-typedef void (*cb_emit_agent_telemetry_t)(char *, char *, float, char *);
+typedef void (*cb_emit_agent_telemetry_t)(char *, char *, double, char *);
 
 // _util
 // (argv, env, stdout, stderr, ret_code, exception)
